Add files via upload
This file contains the code for a Guessing Game, allowing the player to try and guess a randomly generated number. The player can choose between 3 different difficulties providing them with larger number ranges to guess from as the difficulty increases. As a helpful tool when the players guess is one or two off of the randomly generated number the code will tell them they are close by 1 number, if not the computer will let them know they are far off.