Add files via upload
This file codes for a game/quiz that allows the user to determine their Godly parent as if in the universe of Greek Mythology. They can choose between having the parent be a God or a Goddess. They will get a result based on their answer choices which relate to the users personal preferences.