Add files via upload
This file contains the code for a makeshift Pokémon game. It allows the user to use the Pokémon Evee as a their own Pokémon, being able to train, battle and level up their Evee. their progress is also documented like the level changes that occur the more they train or battle.